rsgdfsg

+++ src/input.cpp
@@ -5,7 +5,7 @@ void do_key_up(SDL_KeyboardEvent *event);
 void do_key_down(SDL_KeyboardEvent *event);
 void input(void);
 void check_keys(void);
-bool is_pressed(u8 keybind);
+bool is_pressed(u16 keybind);
 //^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ END Declarations ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 
@@ -32,20 +32,20 @@ void input(void)
 	{
 		switch (event.type)
 		{
-			case SDL_QUIT:
-				exit(0);
-				break;
-				
-			case SDL_KEYDOWN:
-				do_key_down(&event.key);
-				break;
-				
-			case SDL_KEYUP:
-				do_key_up(&event.key);
-				break;
+		case SDL_QUIT:
+			exit(0);
+			break;
+			
+		case SDL_KEYDOWN:
+			do_key_down(&event.key);
+			break;
+			
+		case SDL_KEYUP:
+			do_key_up(&event.key);
+			break;
 
-			default:
-				break;
+		default:
+			break;
 		}
 	}
 	app.mouse.buttons = SDL_GetMouseState(&app.mouse.pos.x, &app.mouse.pos.y);	// get mouse button pos & button state
@@ -57,10 +57,10 @@ void check_keys(void)
 {
 	if(is_pressed(app.keybind.escape))
 		exit(0);
-	
+
 	// player movement
 	app.player.vel = {0.0f, 0.0f};
-	
+
 	if(is_pressed(app.keybind.right))
 	{
 		app.player.vel += move_right();
@@ -88,7 +88,7 @@ void check_keys(void)
 }
 
 // Checks if keybind has been pressed
-bool is_pressed(u8 keybind)
+bool is_pressed(u16 keybind)
 {
 	return (app.keyboard[keybind]) ? true:false;
 }
